+++ src/runtime/index.ts
@@ -1,7 +1,5 @@
-import { EventDataCollector, formatLocation } from '../formatter/helpers'
+import { EventDataCollector } from '../formatter/helpers'
 import StackTraceFilter from '../stack_trace_filter'
-import UserCodeRunner from '../user_code_runner'
-import VError from 'verror'
 import { retriesForPickle, shouldCauseFailure } from './helpers'
 import { IdGenerator } from '@cucumber/messages'
 import * as messages from '@cucumber/messages'
@@ -9,43 +7,13 @@ import TestCaseRunner from './test_case_runner'
 import { EventEmitter } from 'events'
 import { ISupportCodeLibrary } from '../support_code_library_builder/types'
 import TestRunHookDefinition from '../models/test_run_hook_definition'
-import { doesHaveValue, valueOrDefault } from '../value_checker'
 import { ITestRunStopwatch, RealTestRunStopwatch } from './stopwatch'
 import { assembleTestCases } from './assemble_test_cases'
+import { runTestRunHooks } from './run_test_run_hooks'
 
 export interface IRuntime {
   start: () => Promise<boolean>
 }
-
-const runTestRunHooks = (
-  { dryRun }: { dryRun: boolean },
-  { defaultTimeout }: { defaultTimeout: number }
-) =>
-  dryRun
-    ? async () => {}
-    : async (
-        definitions: TestRunHookDefinition[],
-        name: string
-      ): Promise<void> => {
-        for (const hookDefinition of definitions) {
-          const { error } = await UserCodeRunner.run({
-            argsArray: [],
-            fn: hookDefinition.code,
-            thisArg: null,
-            timeoutInMilliseconds: valueOrDefault(
-              hookDefinition.options.timeout,
-              defaultTimeout
-            ),
-          })
-          if (doesHaveValue(error)) {
-            const location = formatLocation(hookDefinition)
-            throw new VError(
-              error,
-              `${name} hook errored, process exiting: ${location}`
-            )
-          }
-        }
-      }
 
 export interface INewRuntimeOptions {
   eventBroadcaster: EventEmitter
